src/django/requirements.txt
Django>=2.1,<3.0
psycopg2>=2.7,<3.0
djangorestframework>=3.9,<4.0